restore BulletEnemy
BulletEnemyを元の状態に戻した。
