--- Assets/Enemies/BulletEnemy/Scripts/BulletEnemyAnimation.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c7a512754efdf9d41a37be9eb2de09ad
