--- Assets/Enemies/BulletEnemy/Scripts/BulletEnemyAttackState.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 6d8bba83112ade14b92c11ba8589225a
