--- Assets/Enemies/BulletEnemy/Scripts/BulletEnemyController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 93c35f3c4aee6e240881324061e3a486
