--- Assets/Enemies/BulletEnemy/Scripts/BulletEnemyDeadState.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f8d50419de4604f419a80cfab66aa04e
