--- Assets/Enemies/BulletEnemy/Scripts/BulletEnemyIdleState.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 55615327050792b40bf3b4bb67fe54d0
