--- Assets/Enemies/BulletEnemy/Scripts/BulletEnemyMoveState.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 6b261b11a293161408eb7ea8fbb73d43
