+++ Assets/Enemies/BulletEnemy/Scripts/EnemyBulletBehaviour.cs
@@ -3,7 +3,6 @@ using UnityEngine;
 public class EnemyBulletBehaviour : MonoBehaviour
 {
     [SerializeField] private float lifeTimeSeconds = 5f;
-    [SerializeField] private float bulletSpeed = 10f;
 
     private StatusActionHolder statusActionHolder;
     private StatusManager statusManager;
@@ -27,10 +26,11 @@ public class EnemyBulletBehaviour : MonoBehaviour
     void Update()
     {
         if (isDestroyed) return;
-        if (rb == null) return;
+        if (rb == null || statusManager == null) return;
 
+        float speed = statusManager.BaseStatus.BaseSpeed;
         Vector2 direction = transform.right;
-        rb.linearVelocity = direction * bulletSpeed;
+        rb.linearVelocity = direction * speed;
     }
 
     void OnTriggerEnter2D(Collider2D collision)
